+++ .env.sample
@@ -1,1 +1,2 @@
-OKTA_API_TOKEN=_GET_FROM_OKTA
+OKTA_ORG_URL=https://{_REPLACE_ME_}.okta.com
+OKTA_API_TOKEN=_REPLACE_ME_
