+++ backend/requirements/common.txt
@@ -22,3 +22,4 @@ twilio==6.9.0
 unittest2==1.1.0
 Werkzeug==0.11.2
 python-dotenv==0.13.0
+polling==0.3.1
